[fix] admin privileges now required to get list of users and delete users

+++ src/routes/userRouter.js
@@ -82,6 +82,11 @@ userRouter.delete(
   authRouter.authenticateToken,
   asyncHandler(async (req, res) => {
     const userId = Number(req.params.userId);
+    const user = req.user;
+    if (!user.isRole(Role.Admin)) {
+      return res.status(403).json({ message: 'unauthorized' });
+    }
+
     await DB.deleteUser(userId);
     res.json({ message: 'user deleted' });
   })
@@ -92,6 +97,10 @@ userRouter.get(
   '/',
   authRouter.authenticateToken,
   asyncHandler(async (req, res) => {
+    const user = req.user;
+    if (!user.isRole(Role.Admin)) {
+      return res.status(403).json({ message: 'unauthorized' });
+    }
     const [users, more] = await DB.getUserList(req.query.page, req.query.limit, req.query.name);
     res.json({ users, more });
   })
